Add autostack support to example
Autostacks created from the CLI do not normally get any automatic
metrics added.  This is deliberate so that expert users do not have
to immediately delete automatic metrics they aren't interested in.

We are explitly adding NetworkIn, DiskWriteBytes and CPUUtilization
metrics to newly created autostacks for the sake of the example.