TAUR-1217 Fix epel repo
The epel repository is available as an RPM. We've been installing it by
just copying a file into `/etc/yum.repos.d`, but that is a bad idea
since that file isn't getting updated and the epel-release RPM is.

The grok-pipeline now uses the `install-epel-repo-6.5` packer script
to remove the old stale repo file, update ca-certificates (if we
don't get current on ca-certificates, the epel repo will cause yum
to fail SSL checks), then install the epel repo via the rpm.

This change was already applied to the grok-plumbing packer
configuration when I initially implemented it for our infrastructure
AMIs in [#72](https://github.com/numenta/numenta-apps/pull/72) somehow
we missed adding it to the grok-pipeline packer configuration then.