TAUR-1324 Remove psutil, supervisord from standard infrastructure load
Instead, let the applications that need them specify the version they
need in their requirements.txt.

If we use a thirdparty app that needs them and _doesn't_ have them in
their requirements.txt we can make a state that installs the version
they need on just the instances using that role.