UNI-187 #comment This commit is fairly stable using `jsondown`.
@lscheinkman @marionleborgne @jcasner FYI! It currently takes TWO clicks on a
Metric to get the effect of saving the data to disk. (The data is actually
being saved to disk correctly on first click, but this remains a mystery
to the DB client for some reason -- this is the issue I'm tracking down.)
Also, `jsondown` is *SLOW* for large operations. (I'll file a separate JIRA
for next phase on migrating away from any batch ops, and finding a final
file-based DB solution).

If you want more speed, comment out `jsondown` and switch to `medeadown`.
Grep for "DatabaseBackend" in `DatabaseServer.js` to switch. Much faster.
Same two-click rule still applies. Harder to debug disk on file. NOT AS STABLE:
The DB can become corrupted easily.

thanks.