DEVOPS-43 Optionally make debug mode less spammy
This allows those of us who don't want to see all that to set
STIFLE_ENVIRONMENT_CONSOLE_SPAM in the environment and have output that
is easier to read if you're not changing the environment between
running external commands, even when the log level is set to debug.

I almost never want to see the environment when I'm running a script
during development unless I'm modifying the environment that I'm passing
to external commands that I'm calling. That said, when scripts are running
in long run-time jenkins job we decided we do want to dump all the
environment variables to the console so the jobs don't need to be re-run
with special flags to see them.