UNI-328 #comment **Very rough** Proof of concept moving the Anomaly Bar Charts away from being a Dygraph Series Plotter (core to Dygraphs), towards being a Dygraph Plugin (more general, less easy).
This is because Dygraphs only supports 2 Y-Axes, and we need 3 (value, anomaly
bars, non-aggregated value). So, I'm trying to reserve the 2 y-axes for the line
charts (value, non-agg value), and moving anomaly bars to a custom plugin
(instead of the nice easy series plotter which was already working).  I think I
can get this back up to spec with +1 more day of work.

I'm going to pause the Plugin approach temporarily, however. I just found
DyGraph's "Underlay Callback" feature:
  http://dygraphs.com/tests/underlay-callback.html
This may be a better approach than the Plugin, so I'm going to explore it
next for a bit, first.

thanks. FYI @marionleborgne @lscheinkman @cbaranski